+++ src/app/feature/home/components/card/card.component.ts
@@ -1,6 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { Character } from '@home/shared/model/character.model';
 import { CharacterService } from '@home/shared/services/character/character.service';
+import { ModalComponent } from '@shared/components/modal/modal.component';
 
 @Component({
   selector: 'app-card',
@@ -11,9 +12,10 @@ export class CardComponent implements OnInit {
 
   characters!: Character[];
   record!: Character;
+  @ViewChild(ModalComponent) modal!: ModalComponent;
 
   constructor(
-    private apiCharacter: CharacterService
+    private apiCharacter: CharacterService,
   ) { }
 
   ngOnInit(): void {
@@ -27,8 +29,13 @@ export class CardComponent implements OnInit {
     });
   }
 
-  getOne() {
-
+  getOne(id:number) {
+    this.apiCharacter.getOne(id).subscribe(response => {
+      const [character] = response?.data?.results;
+      this.record = character;
+      console.log(this.record);
+      this.modal.open(this.record, false);
+    });
   }
 
 }