+++ src/app/shared/components/modal/modal.component.ts
@@ -7,18 +7,23 @@ import { Component, Input, OnInit, Output, EventEmitter } from '@angular/core';
 })
 export class ModalComponent implements OnInit {
 
-  @Input() record!: any;
-  @Input() isComic!: boolean;
+  record!: any;
+  isComic!: boolean;
   @Output() data? = new EventEmitter();
   modal: boolean = false;
+  image!: string;
 
   constructor() { }
 
   ngOnInit(): void {
   }
 
-  open() {
+  open(data: any, isComic: boolean) {
     this.modal = true;
+    this.isComic = isComic;
+    this.record = data;
+    this.image = `${this.record?.thumbnail?.path}/portrait_incredible.${this.record?.thumbnail?.extension}`;
+    console.log(this.record);
   }
 
   close() {